chore: cleans up root layout from dependency checker
